+++ src/mpc/inputs.jl
@@ -20,6 +20,8 @@ struct MPCInputs <: AbstractInputs
     ratchets::UnitRange
     techs_by_exportbin::DenseAxisArray{Array{String,1}}  # indexed on [:NEM, :WHL]
     export_bins_by_tech::Dict{String, Array{Symbol, 1}}
+    storage_by_exportbin::DenseAxisArray{Array{String,1}}  # indexed on [:NEM, :WHL]
+    export_bins_by_storage::Dict{String, Array{Symbol, 1}} 
     cooling_cop::Dict{String, Array{Float64,1}}  # (techs.cooling, time_steps)
     thermal_cop::Dict{String, Float64}  # (techs.absorption_chiller)
     ghp_options::UnitRange{Int64}  # Range of the number of GHP options
@@ -64,6 +66,7 @@ function MPCInputs(s::MPCScenario)
         export_bins_by_tech[t] = s.electric_tariff.export_bins
     end
     # TODO implement export bins by tech (rather than assuming that all techs share the export_bins)
+    # TODO: account for storage export but only if Storage can_net_meter etc are true 
  
     #Placeholder COP because the REopt model expects it
     cooling_cop = Dict("ExistingChiller" => ones(length(s.electric_load.loads_kw)) .* s.cooling_load.cop)
@@ -92,6 +95,8 @@ function MPCInputs(s::MPCScenario)
         1:length(s.electric_tariff.tou_demand_ratchet_time_steps),  # ratchets
         techs_by_exportbin,
         export_bins_by_tech,
+        storage_by_exportbin,
+        export_bins_by_storage,
         cooling_cop,
         thermal_cop,
         ghp_options,